ipv6: Remove external dependency on rt6i_dst and rt6i_src
This patch removes the assumptions that the returned rt is always
a RTF_CACHE entry with the rt6i_dst and rt6i_src containing the
destination and source address.   The dst and src can be recovered from the
calling site.

We may consider to rename (rt6i_dst, rt6i_src) to (rt6i_key_dst, rt6i_key_src)
later. They should be used as the keys to the fib6 tree only.

Signed-off-by: Martin KaFai Lau <kafai@fb.com>