ipv6: Extend the route lookups to low priority metrics.
Before patch 'Allow pmtu update on /128 via gateway route',
RTF_CACHE route was not created for DST_HOST routing via a gateway and
it was a bug because a pmtu exception will cause the DST_HOST route to expire.
The bug is now fixed but it also requires changes on both delete code path
and rt6_select() code path.

This patch is to fix the rt6_select() code path by extending the selection
to routes with lower priority metrics if the higher priority route(s) is
invalid (like expired).

We search only for routes with highest priority metric in
find_rr_leaf(). However if one of these routes is marked
as invalid, we may fail to find a route even if there is
a appropriate route with lower priority. Then we loose
connectivity until the garbage collector deletes the
invalid route. This typically happens if a host route
expires afer a pmtu event. Fix this by searching also
for routes with a lower priority metric.

Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Martin KaFai Lau <kafai@fb.com>