ipv6: Remove external dependency on rt6i_gateway and RTF_ANYCAST
When creating a RTF_CACHE route, RTF_ANYCAST is set accordingly based on
rt6i_dst.  Also, rt6i_gateway is always set to the nexthop while the nexthop
could be a gateway or the rt6i_dst.addr.

After removing the rt6i_dst and rt6i_src dependency in the last patch, we also
need to stop the caller from depending on rt6i_gateway and RTF_ANYCAST.

Signed-off-by: Martin KaFai Lau <kafai@fb.com>