ipv6: Break up ip6_rt_copy()
This patch breaks up ip6_rt_copy() into ip6_rt_copy_init() and
ip6_rt_cache_copy().

In the later patch, after seeing pmtu exceptions, we need to invalidate
sockets (like TCP) which have cached the rt.  To do this,
we need to make a copy of the cached rt and reinsert it to the fib6 tree.
This patch refactors the common rt init codes into ip6_rt_copy_init()
which can be reused in the re-insertion logic later.

Signed-off-by: Martin KaFai Lau <kafai@fb.com>