ipv6: Remove rt6_set_from() and its dst.from usage
This patch removes the dst.from chain which is used to track the
the expiration of the original route.  A ref-counted struct rt6_shinfo is
used to replace it.

It is a prep step to the later patch
'Only create RTF_CACHE routes after encountering exceptions'.  When we
create RTF_CACHE route after seeing pmtu update, we want to do the followings
in order to invalidate everyone cached the original route (like TCP):
1. create a new original route
2. delete the old original route from fib6 tree
3. insert the new original route to fib6 tree

The dst.from chain makes this re-insertion difficult.

Signed-off-by: Martin KaFai Lau <kafai@fb.com>