ipv6: Allow pmtu update on /128 via gateway route
Consider there is a permanent /128 via gateway route (DST_HOST) in
the route table.  When there is a pmtu update, the pmtu DST_HOST route is
updated and the RTF_EXPIRES is set.  The permanent DST_HOST route will be
removed after expiration.

The patch is to create a RTF_CACHE route even for the DST_HOST route.  This
change will break some cases, like route deletion, route selection and
dst metrics update.  They will be addressed in the following patches.

Since we are at it, the patch is trying to simplify some checking cases in
ip6_rt_update_pmtu().

1. !(rt6->rt6i_flags & RTF_CACHE) is used to decide when
a RTF_CACHE route needs to be created for pmtu update.  I think a RTF_CACHE
route should be created whenever there is a pmtu update.  It is a bold
change and needs more eyes.

We can also keep using the old conditions and add another RTF_GATEWAY check.

2. Remove the rt6->rt6i_dst.plen == 128 check since RTF_CACHE route will
be created (if it is needed) before updating the mtu.

3. Add a check to ensure no pmtu update on RTF_LOCAL route

Signed-off-by: Martin KaFai Lau <kafai@fb.com>