ipv6: Consider RTF_CACHE when searching the fib6 tree
It is a prep work for the later bug-fix patch which will stop /128 route
from disappearing after pmtu update.

The later bug-fix patch will allow a /128 route and its RTF_CACHE both
exist on the same fib6_node.  To do this, we need to prepare the
existing fib6 tree search to expect RTF_CACHE and decide
which one (clone or non-clone) to pick.

Additional change:
1. Add a missing (!fn) check to __ip6_route_redirect().

Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Cc: Steffen Klassert <steffen.klassert@secunet.com>